+++ src/Graphics/DepthFunction.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace osc
+{
+    enum class DepthFunction
+    {
+        Less = 0,
+        LessOrEqual,
+        TOTAL,
+        Default = Less,
+    };
+}