+++ src/Graphics/TextureFormat.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <cstdint>
+
+namespace osc
+{
+    enum class TextureFormat : int32_t {
+        RGBA32 = 0,
+        RGB24,
+        TOTAL
+    };
+}